Commit before chat and video call

+++ Backend/src/controllers/cart.controller.ts
@@ -1,6 +1,9 @@
 import { Request, Response, NextFunction } from "express";
 import { CourseService } from "../services/course.service";
 import { decodedToken } from "../helperFunction/authHelper";
+import { HttpException } from "../middleware/error.middleware";
+import STATUS_CODES from "../constants/statusCode";
+import MESSAGES from "../constants/message";
 
 
 export class CartController {
@@ -17,16 +20,16 @@ export class CartController {
     ): Promise<void> => {
         try {
             const Token = req.cookies.user
-            console.log(Token)
             const token = Token.accessToken;
+            if(!token){
+                throw new HttpException(STATUS_CODES.UNAUTHORIZED,MESSAGES.ERROR.UNAUTHORIZED)
+            }
             const requiredRole = "user";
             const userId: string | null = decodedToken(token, requiredRole);
             const { courseId } = req.body;
-            console.log("UserId", userId, "courseId:", courseId)
             const response = await this.courseService.addToCart(userId, courseId);
             res.status(201).json({
-                message: "Cart item fetched successfully",
-                Data: "data"
+                message: response,
             });
 
         } catch (error) {
@@ -41,20 +44,23 @@ export class CartController {
     ): Promise<void> => {
         try {
             const Token = req.cookies.user
-            console.log(Token)
-
             const token = Token.accessToken;
             const requiredRole = "user";
+            if(!token){
+                throw new HttpException(STATUS_CODES.UNAUTHORIZED,MESSAGES.ERROR.UNAUTHORIZED)
+            }
             const userId: string | null = decodedToken(token, requiredRole);
             const items = await this.courseService.getCartItems(userId);
-            console.log(items, "Items")
+            if (!items) {
+                throw new HttpException(STATUS_CODES.NOT_FOUND, MESSAGES.ERROR.DATA_NOTFOUND)
+              }
             res.status(201).json({
                 message: "Cart item fetched successfully",
                 Data: items
             });
 
         } catch (error) {
-            next(error);
+            next(error)
         }
     }
 
@@ -63,16 +69,16 @@ export class CartController {
         res: Response,
         next: NextFunction
     ): Promise<void> => {
-        try {
-            console.log("Remove request initiated");
-            
+        try {            
             const Token = req.cookies.user
             const token = Token.accessToken;
+            if(!token){
+                throw new HttpException(STATUS_CODES.UNAUTHORIZED,MESSAGES.ERROR.UNAUTHORIZED)
+            }
             const requiredRole = "user";
             const userId: string | null = decodedToken(token, requiredRole);
             const { courseId } = req.body;
-            console.log("UserId", userId, "courseId:", courseId)
-            const response = await this.courseService.removeCart(userId, courseId);
+            await this.courseService.removeCart(userId, courseId);
             res.status(201).json({
                 message: "Cart item removed successfully",
             });
